Add initial binbuf support, not used yet
Add binbuf_init() and binbuf_free() and create binbuf.c and binbuf.h,
not used yet. Provides `struct binbuf`:

```c
struct binbuf {
        size_t alloc;
        size_t len;
        char *buf;
};
```

In addition to the allocated buffer, it also stores the size of the
allocated memory, and length of the buffer. This means it can store
binary data and text containing null bytes ('\0').

9b20cada-9eaa-11ef-8c04-83850402c3ce