selftest.c: Create `failcount` file-static variable, updated by ok()
Let ok() update the variable so it's always in sync with the number of
"not ok" lines. It's not necessary to increment local variables returned
to a parent function any longer. Sometimes ok() is used in places where
there's not a `return` mechanism to propagate the error count all the
way up to opt_selftest().

9852db10-fc4a-11ef-9056-83850402c3ce