std_strerror(): Write "Unknown errno" message on a new line
Because std_strerror() complains about unknown errno during output from
myerror(), both messages got merged. Insert a newline before the message
from std_strerror() so it looks a bit better. Not perfect, but good
enough.

And while we're in the area, refer to the value as "errno" instead of
"errnum" to avoid confusion.

6a116086-2964-11f0-b11b-83850402c3ce