Fix 3 failing tests on FreeBSD 11.0
These tests started to fail in commit `3e0bfd76a2ac` ("std.c: myerror()
resets errno if it contains an error value", 2025-05-05):

- not ok 571 - Empty argument to --count (stderr)
- not ok 574 - --count receives non-number (stderr)
- not ok 664 - Empty argument to --seed (stderr)

When strtol(3) on FreeBSD receives an empty string or invalid number, it
sets errno to EINVAL, and myerror() prints out an additional ": Invalid
argument\n" string. According to the man page for strtol(3) this
behavior is not portable on all platforms, so add 2 preprocessor
directives in choose_opt_action() that set errno to 0 if it's EINVAL and
`endptr` points to `optarg` or `dest->seed`.

d5ed6ff2-365c-11f0-8486-f7f8c6927076