Makefile: longlines: Switch to Perl for consistent UTF-8 handling
Commit `12b80714e146` ("Fix `longlines` error on OpenBSD 7.7, use awk(1)
instead of grep(1)", 2025-06-16) resolved the issue on OpenBSD but broke
`make longlines` on Debian-based systems (e.g., Devuan). The default
`mawk` there doesn't handle UTF-8 well, requiring `gawk` to be installed
instead.

For context, the check fails on line 112 in geomath.c:

```
$ make longlines
geomath.c
112:  * For exact pole positions (lat = ±90°), the latitude is adjusted by ≈1 cm to

make: *** [Makefile:248: longlines] Error 1
$
```

Rewording the line to avoid the problem could be an option, but it's
2025 and a matter of principle at this point. UTF-8 has been the
dominant encoding for around 20 years, it's about time the software
adapts.

So, because we don't know if the installed `grep` or `awk` is retarded,
use Perl instead. It's everywhere and handles UTF-8 correctly.

5f36296e-4d70-11f0-baee-83850402c3ce