selftest.c: Make `execname` a file-static variable
Move `execname` out of opt_selftest() to the file scope, and declare it
as a file-static variable initialized in opt_selftest().

In a future commit test_command() will need the value of `argv[0]`,
provided by `execname`. This means all calls to tc() and sc() would need
to pass `execname` as a parameter, and that's cumbersome and painful.

As a bonus, this change also gets rid of many `execname` parameters in
other functions.

878d7398-5484-11f0-bcbc-83850402c3ce